rename: 쿠폰 컴포넌트 경로 변경 (#95)
