+++ fe/src/components/coupon/common/CouponGroup.style.ts
