+++ fe/src/components/coupon/common/CouponGroupCard.tsx
@@ -1,11 +1,11 @@
 /** @jsxImportSource @emotion/react */
 import { cardStyle, inputBoxGropupFooterStyle } from './CouponGroup.style';
-import CouponRegistrationGroup from './registrationGroup/CouponRegistrationGroup';
-import DateInputGroup from './inputGroup/DateInputGroup';
-import TextInputGroup from './inputGroup/TextInputGroup';
+import CouponRegistrationGroup from '../registrationGroup/CouponRegistrationGroup';
+import DateInputGroup from '../inputGroup/DateInputGroup';
+import TextInputGroup from '../inputGroup/TextInputGroup';
 import { useInput } from '@hooks/useInput';
-import { useCoupon } from '@components/coupon/useCoupons';
-import SelectInputGroup from './inputGroup/SelectInputGroup';
+import { useCoupon } from '@components/coupon/common/useCoupons';
+import SelectInputGroup from '../inputGroup/SelectInputGroup';
 import CustomButton from '@components/common/CustomButton/CustomButton';
 import { Icon } from '@components/common/Icon/Icon';
 import { CouponGroupType, IsRandom } from './CouponType';
